refactor: Move CreateTemplate() and have 1 decoderFunc
